Introduce InstallUndoContext

+++ overlord/snapstate/backend/setup.go
@@ -27,15 +27,16 @@ import (
 	"github.com/snapcore/snapd/boot"
 	"github.com/snapcore/snapd/progress"
 	"github.com/snapcore/snapd/snap"
+	"github.com/snapcore/snapd/snap/undo_context"
 )
 
 // SetupSnap does prepare and mount the snap for further processing.
-func (b Backend) SetupSnap(snapFilePath, instanceName string, sideInfo *snap.SideInfo, meter progress.Meter) (snapType snap.Type, err error) {
+func (b Backend) SetupSnap(snapFilePath, instanceName string, sideInfo *snap.SideInfo, meter progress.Meter) (snapType snap.Type, undoCtx *undo_context.InstallUndoContext, err error) {
 	// This assumes that the snap was already verified or --dangerous was used.
 
 	s, snapf, oErr := OpenSnapFile(snapFilePath, sideInfo)
 	if oErr != nil {
-		return snapType, oErr
+		return snapType, nil, oErr
 	}
 
 	// update instance key to what was requested
@@ -47,43 +48,45 @@ func (b Backend) SetupSnap(snapFilePath, instanceName string, sideInfo *snap.Sid
 		if err == nil {
 			return
 		}
+
 		// XXX: this will also remove the snap from /var/lib/snapd/snaps
-		if e := b.RemoveSnapFiles(s, s.GetType(), meter); e != nil {
+		if e := b.RemoveSnapFiles(s, s.GetType(), undoCtx, meter); e != nil {
 			meter.Notify(fmt.Sprintf("while trying to clean up due to previous failure: %v", e))
 		}
 	}()
 
 	if err := os.MkdirAll(instdir, 0755); err != nil {
-		return snapType, err
+		return snapType, nil, err
 	}
 
 	if s.InstanceKey != "" {
 		err := os.MkdirAll(snap.BaseDir(s.SnapName()), 0755)
 		if err != nil && !os.IsExist(err) {
-			return snapType, err
+			return snapType, nil, err
 		}
 	}
 
-	if err := snapf.Install(s.MountFile(), instdir); err != nil {
-		return snapType, err
+	if undoCtx, err = snapf.Install(s.MountFile(), instdir); err != nil {
+		return snapType, nil, err
 	}
 
 	// generate the mount unit for the squashfs
 	if err := addMountUnit(s, meter); err != nil {
-		return snapType, err
+		return snapType, nil, err
 	}
 
 	if s.GetType() == snap.TypeKernel {
 		if err := boot.ExtractKernelAssets(s, snapf); err != nil {
-			return snapType, fmt.Errorf("cannot install kernel: %s", err)
+			return snapType, nil, fmt.Errorf("cannot install kernel: %s", err)
 		}
 	}
 
-	return s.GetType(), err
+	// non-nil undoCtx is only returned on success, as failures are handled immediately by the deferred cleanup of this function.
+	return s.GetType(), undoCtx, err
 }
 
 // RemoveSnapFiles removes the snap files from the disk after unmounting the snap.
-func (b Backend) RemoveSnapFiles(s snap.PlaceInfo, typ snap.Type, meter progress.Meter) error {
+func (b Backend) RemoveSnapFiles(s snap.PlaceInfo, typ snap.Type, undoCtx *undo_context.InstallUndoContext, meter progress.Meter) error {
 	mountDir := s.MountDir()
 
 	// this also ensures that the mount unit stops
@@ -105,9 +108,12 @@ func (b Backend) RemoveSnapFiles(s snap.PlaceInfo, typ snap.Type, meter progress
 			}
 		}
 
-		// remove the snap
-		if err := os.RemoveAll(snapPath); err != nil {
-			return err
+		// don't remove snap path if it existed before snap installation was attempted
+		if undoCtx == nil || !undoCtx.TargetPathExists {
+			// remove the snap
+			if err := os.RemoveAll(snapPath); err != nil {
+				return err
+			}
 		}
 	}
 
@@ -132,6 +138,6 @@ func (b Backend) RemoveSnapDir(s snap.PlaceInfo, hasOtherInstances bool) error {
 }
 
 // UndoSetupSnap undoes the work of SetupSnap using RemoveSnapFiles.
-func (b Backend) UndoSetupSnap(s snap.PlaceInfo, typ snap.Type, meter progress.Meter) error {
-	return b.RemoveSnapFiles(s, typ, meter)
+func (b Backend) UndoSetupSnap(s snap.PlaceInfo, typ snap.Type, undoCtx *undo_context.InstallUndoContext, meter progress.Meter) error {
+	return b.RemoveSnapFiles(s, typ, undoCtx, meter)
 }
